Work was furthered on the Engine. There are still crucial player commands that are missing, but a simple game can exist with the current engine.
The addition of the Engine class has simplified world creation dramatically. The creation of rooms and connections, looks much more clean and concise, letting the Engine handle all the work of putting together the world, not the programmer/creator.

I may want to consider creating an editor program that people can use to create and edit their adventures/mazes/puzzles.

I would like to add puzzle support and usable item support for this as well. This is a possibility that I haven't figured out yet, but I think would be highly beneficial to the overall game.